+++ R/plot.R
@@ -3,16 +3,15 @@ NULL
 
 setGeneric("plot")
 
-#' Plot method for occumbFit class.
-#'
-#' @param x An occumbFit object.
-#' @param y Not used.
-#' @param ... Other arguments passed to the plot method for
-#'            \code{jagsUI} class.
-#' @export
 setMethod("plot", signature(x = "occumbFit"),
     function(x, ...) {
         plot(x@fit, ...)
+    }
+)
+
+setMethod("plot", signature(x = "occumbGof"),
+    function(x, ...) {
+        plot_gof(x@stats_obs, x@stats_rep, x@stats)
     }
 )